Updated downscaling
My laptop crapped out on me in the middle of the night, so the downscaling results still haven't been QC'd. I'm rerunning now, but it should be done within an hour. I'll look through the results and forward them to you if everything looks ok. Best guess right now is that it takes about an hour and a half to run on my low-end laptop. I'm sure there's lots of room to speed that up, even if it doesn't happen for this class project.